Hit and Hurt boxes working
Hit and Hurt boxes are working.

Parry has been discontinued due to the time contrants. there is no way we can finish it before the deadline.

There exists a quite high possibility that the way the animations and states in the Finate State Machine are working is gonna need to be reworked in order to allow animating the player and enemy sprites. Unless I can find some way to make it work like this. It's gonna be clunky, but that is all we have time for before the deadline, so it's gonna have to be clunky