published version for the game jam
it is "finished" not FINISHED, but "finished". it is janky, there is no sounds, some of the sprites havent been made, it is buggy as heck and you can get stuck in walls. BUT IT IS DONE AND SUBMITTED 30 SECONDS BEFORE THE DEADLINE!! so im counting this as a successful learning experience